requirements.txt
Flask>=2.0,<3.0
spotipy>=2.22.1
python-dotenv>=1.0
